COCALC: Only upgrade python3 version of mercurial.
Currently, the default system version of mercurial on cocalc
is python2 based.  This does not have proper development libraries
and so it fails to upgrade mercurial.  We could fallback to the old
approach of just upgrading hg-git and evolve, but I think it is better
just to install the python3 version of mercurial, so I do only that now.

--HG--
branch : 0.3.1